Fix WebXR stencil test
According to the WebGL spec, about the readPixel() function:
"Only two combinations of format and type are accepted. The
 first is format RGBA and type UNSIGNED_BYTE. The second is
 an implementation-chosen format. The values of format and
 type for this format may be determined by calling
 getParameter with the symbolic constants
 IMPLEMENTATION_COLOR_READ_FORMAT and
 IMPLEMENTATION_COLOR_READ_TYPE, respectively. The
 implementation-chosen format may vary depending on the
 format of the currently bound rendering surface.
 Unsupported combinations of format and type will generate
 an INVALID_OPERATION error."

This means using GL_RGB is platform dependent, while using
GL_RGBA is supported by all platforms, so the test was
modified to use GL_RGBA in order to be spec compliant.

Bug: chromium:1204234
Change-Id: Icfc9236f5281f5bc9d396b1dce1a86978facf8c8