kurl: remove tabs and newlines from setter inputs
This makes setters of Location.search, HTMLAnchorElement.hash,
URL.pathname, etc., remove tabs and newlines. This aligns with the
WHATWG URL Standard and Safari, as well as Firefox partially.

There are a few exceptions. SetProtocol is unchanged due to an upstream
spec issue [1]. SetUser and SetPass should not remove tabs and newlines
per spec.

[1]: https://github.com/whatwg/url/issues/609

Bug: 1214932
Change-Id: I3bcb6fa9e3b61ee5a0a03b13eca93d42912afb22