Add tests to verify the layout shift score stay unchanged for touch drag
Our current implementation in LayoutShiftTracker::NotifyInput puts all
the events after a Pointerdown event in a queue, and keep accumulating
the layout shift score, once we see a PointerUp, we set the PointerUp
event to be hasRecentInput, so the layout shift score will not increase
for this touch dragging case. But for touch scroll, we will send the
PointerCancel event instead, so the accumulating shift score will be
reported.

Bug: 1166952
Change-Id: Ia4b25d894b5787fdc7f244ab9a8006bc3ff69cff